Refactor oligo property calculation module
- Removed the OligoAttributes class from the oligo_designer_toolsuite.database module.
- Introduced a new oligo_property_calculator module with various property calculation classes and functions.
- Added classes for calculating properties such as length, GC content, melting temperature, and more.
- Implemented parallel processing for property calculations across oligonucleotides in the OligoDatabase.
- Updated the structure to enhance modularity and maintainability of property calculations.
