+++ oligo_designer_toolsuite/database/__init__.py
@@ -3,13 +3,11 @@ This module provides key classes for managing and processing oligonucleotide-rel
 """
 
 from ._oligo_database import OligoDatabase
-from ._oligo_database_attributes import OligoAttributes
 from ._reference_database import ReferenceDatabase
 
 __all__ = [
     "OligoDatabase",
     "ReferenceDatabase",
-    "OligoAttributes",
 ]
 
 classes = __all__
